Add True Class Name
Class name and file name were being conflated. In most cases the two are
the same but not always.

I renamed the existing className property in RankedDisharmony to
fileName to reflect what it actually is.

I added a new className property which is set to the actual
className.

To get class name passed into the logic where RankedDisharmony is
constructed I also had to update the GodClass definition to include
className as well as file name.

The html and cvs report output should not change, as I updated them to
call getFileName from ranked disharmony rather than getClassName. This
preserves existing behavior.

The only output which should change is the  json report which will now expose both fileName and className, letting the
consumer choose which they care about.